Refactoring actions

+++ src/main/java/com/mtrifonov/springsecuritystandard/services/RoleChangeService.java
@@ -1,7 +1,13 @@
 package com.mtrifonov.springsecuritystandard.services;
 
+import com.mtrifonov.springsecuritystandard.Role;
 import com.mtrifonov.springsecuritystandard.repositories.RoleRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+
+import lombok.AllArgsConstructor;
+
+import java.util.List;
+
+
 import org.springframework.stereotype.Service;
 
 /**
@@ -9,26 +15,25 @@ import org.springframework.stereotype.Service;
  * @Mikhail Trifonov
  */
 @Service
+@AllArgsConstructor
 public class RoleChangeService {
     
     private final RoleRepository roleRepo;
-
-    @Autowired
-    public RoleChangeService(RoleRepository roleRepo) {
-        this.roleRepo = roleRepo;
+    
+    public void changeRoles(List<Role> rolesForAdd, List<Role> rolesForDelete, int id) {
+        
+        if (rolesForAdd.size() != 0) {
+            addRoles(rolesForAdd, id);
+        } else {
+            deleteRoles(rolesForDelete, id);
+        }
     }
     
-    private void addRoles(String[] roles, int id) {
+    private void addRoles(List<Role> roles, int id) {
         roleRepo.addRolesById(roles, id);
     }
     
-    private void deleteRoles(String[] roles, int id) {
+    private void deleteRoles(List<Role> roles, int id) {
         roleRepo.deleteRolesById(roles, id);
     }
-
-    public void changeRoles(String[] rolesForAdd, String[] rolesForDelete, int id) {
-        addRoles(rolesForAdd, id);
-        deleteRoles(rolesForDelete, id);
-    }
-    
 }
